move kcfg.minimize into a separate file and class
